Cast audit to the expected class on load from database
The Durable module exposes an `auditor` attribute that provides for
assigning a customized auditor to a job class in place of the
standard QueueAudit.

When a job is reenqueued from the database, the QueueAudit class
pushes an instance of itself onto the arguments list and calls
`enqueue` on the job class. In `enqueue`, in turn, the Durable
module looks for an instance of the job class's auditor class as
the last argument in the list. If it finds one, it is popped from
the list and its `enqueued_id` is pushed onto the list. If not, a
new audit is construct and its `enqueued_id` is pushed onto the list.

Currently, because the audit restored from the database is
unconditionally constructed as a QueueAudit, Durable cannot recognize
it as the audit it needs when the job class has an alternative
auditor defined (such as a subclass of QueueAudit). This change
resolves this issue by casting the audit type to the job class's
auditor type before pushing it onto the job arguments list.