[CCORESF2-1501] Don't use Thread#signal in heartbeat implementation
This was working fine if the heartbeat thread was blocked in sleep, but
if the heartbeat thread was in the process of calling MySQL it would
send a SIGVTALRM to it (because the unblocking function is invoked).
This had various undesirable side effects inside the mysql2 gem native
extension, including rapid-fire disconnect/reconnect cycles every 10ms.

Instead, use a condition variable that the heartbeat thread will be
blocked on during the "sleep" phase of the heartbeat. Then, the condvar
will be signaled when canceling the heartbeat thread. This means that
the heartbeat thread is interruptable during sleep, but will finish any
pending call to mysql2 before exiting.