add possibility to disable immediate requeuing
A test that enabled `requeue_immediately` in `durable_test` caused flakiness in tests elsewhere: the 'does not get tried again' test in `resque_test` was failing if the `requeue_immediately` ran before it. This happened because after enabling `requeue_immediately` there was no way to reset it, and instead of being completed a job in the second spec was immediately rescheduled.